service(notification): show toast on new notification
