+++ src/app/navbar/navbar-notifications/navbar-notifications.component.ts
@@ -8,6 +8,7 @@ import { filter, map, mergeMap, share } from 'rxjs/internal/operators';
 import { SeenNotification } from '../../services/notification/notification.actions';
 import * as moment from 'moment';
 import { NOTIFICATION_TYPE } from '../../services/notification/notification.constant';
+import { NotificationService } from '../../services/notification/notification.service';
 
 @Component({
   selector: 'app-navbar-notifications',
@@ -25,7 +26,7 @@ export class NavbarNotificationsComponent implements OnInit {
 
   constructor(
     private store: Store<IAppState>,
-    private router: Router,
+    public notificationService: NotificationService,
   ) {
     this.notifications$ = this.store.pipe(select('notification', 'list'));
   }
@@ -68,36 +69,6 @@ export class NavbarNotificationsComponent implements OnInit {
 
   toggleFocus() {
     this.notificationsToggle = true;
-  }
-
-  getSenderImage(notification: DNotification): string {
-    if (notification.senderCharacter) return notification.senderCharacter.portrait.px256x256;
-    if (notification.senderCorporation) return notification.senderCorporation.icon.px256x256;
-    if (notification.senderAlliance) return notification.senderAlliance.icon.px128x128;
-  }
-
-  getSenderName(notification: DNotification): string {
-    if (notification.senderCharacter) return notification.senderCharacter.name;
-    if (notification.senderCorporation) return notification.senderCorporation.name;
-    if (notification.senderAlliance) return notification.senderAlliance.name;
-  }
-
-  getText(notification: DNotification): string {
-    const name = this.getSenderName(notification);
-    switch (notification.type) {
-      case NOTIFICATION_TYPE.NEW_POST_ON_YOUR_WALL:
-        return `<strong>${name}</strong> has posted on <strong>your wall</strong>.`;
-      case NOTIFICATION_TYPE.NEW_POST_ON_YOUR_CORPORATION_WALL:
-        return `<strong>${name}</strong> has posted on <strong>your corporation wall</strong>.`;
-      case NOTIFICATION_TYPE.NEW_POST_ON_YOUR_ALLIANCE_WALL:
-        return `<strong>${name}</strong> has posted on <strong>your alliance wall</strong>.`;
-      case NOTIFICATION_TYPE.NEW_COMMENT_ON_YOUR_POST:
-        return `<strong>${name}</strong> has commented on <strong>your post</strong>.`;
-      case NOTIFICATION_TYPE.NEW_COMMENT_ON_A_POST_YOU_PARTICIPATE:
-        return `<strong>${name}</strong> has commented on <strong>a post you participate</strong>.`;
-      default:
-        return `Unknown notification type: ${notification.type}`;
-    }
   }
 
 }
